nova 14.0.0.0b2 development milestone

meta:version: 14.0.0.0b2
meta:series: newton
meta:release-type: development milestone
meta:announce: openstack-announce@lists.openstack.org
meta:pypi: no
meta:first: no
meta:release:Author: Matt Riedemann <mriedem@us.ibm.com>
meta:release:Commit: Matt Riedemann <mriedem@us.ibm.com>
meta:release:Change-Id: If0c0fba5f11f9e366e6cc2047dcad18d5b37c309
meta:release:Code-Review+2: Doug Hellmann <doug@doughellmann.com>
meta:release:Workflow+1: Doug Hellmann <doug@doughellmann.com>

+++ nova/virt/xenapi/vm_utils.py
@@ -58,6 +58,7 @@ from nova.virt.image import model as imgmodel
 from nova.virt import netutils
 from nova.virt.xenapi import agent
 from nova.virt.xenapi.image import utils as image_utils
+from nova.virt.xenapi import volume_utils
 
 LOG = logging.getLogger(__name__)
 
@@ -893,12 +894,15 @@ def _auto_configure_disk(session, vdi_ref, new_gb):
         2. The disk must have only one partition.
 
         3. The file-system on the one partition must be ext3 or ext4.
+
+        4. We are not running in independent_compute mode (checked by
+           vdi_attached)
     """
     if new_gb == 0:
         LOG.debug("Skipping auto_config_disk as destination size is 0GB")
         return
 
-    with vdi_attached_here(session, vdi_ref, read_only=False) as dev:
+    with vdi_attached(session, vdi_ref, read_only=False) as dev:
         partitions = _get_partitions(dev)
 
         if len(partitions) != 1:
@@ -920,6 +924,10 @@ def _auto_configure_disk(session, vdi_ref, new_gb):
 
 
 def try_auto_configure_disk(session, vdi_ref, new_gb):
+    if CONF.xenserver.independent_compute:
+        raise exception.NotSupportedWithOption(
+            operation='auto_configure_disk',
+            option='CONF.xenserver.independent_compute')
     try:
         _auto_configure_disk(session, vdi_ref, new_gb)
     except exception.CannotResizeDisk as e:
@@ -963,13 +971,20 @@ def _generate_disk(session, instance, vm_ref, userdevice, name_label,
                    disk_type, size_mb, fs_type, fs_label=None):
     """Steps to programmatically generate a disk:
 
-        1. Create VDI of desired size
+    1. Create VDI of desired size
 
-        2. Attach VDI to compute worker
+    2. Attach VDI to Dom0
 
-        3. Create partition
+    3. Create partition
+    3.a. If the partition type is supported by dom0 (currently ext3,
+    swap) then create it while the VDI is attached to dom0.
+    3.b. If the partition type is not supported by dom0, attach the
+    VDI to the domU and create there.
+    This split between DomU/Dom0 ensures that we can create most
+    VM types in the "isolated compute" case.
 
-        4. Create VBD between instance VM and VDI
+    4. Create VBD between instance VM and VDI
+
     """
     # 1. Create VDI
     sr_ref = safe_find_sr(session)
@@ -979,16 +994,26 @@ def _generate_disk(session, instance, vm_ref, userdevice, name_label,
                          virtual_size)
 
     try:
-        # 2. Attach VDI to compute worker (VBD hotplug)
-        with vdi_attached_here(session, vdi_ref, read_only=False) as dev:
+        # 2. Attach VDI to Dom0 (VBD hotplug)
+        mkfs_in_dom0 = fs_type in ('ext3', 'swap')
+        with vdi_attached(session, vdi_ref, read_only=False,
+                               dom0=True) as dev:
             # 3. Create partition
-            partition_start = "2048s"
-            partition_end = "-0"
+            partition_start = "2048"
+            partition_end = "-"
 
-            partition_path = _make_partition(session, dev,
-                                             partition_start, partition_end)
+            session.call_plugin_serialized('partition_utils.py',
+                                           'make_partition', dev,
+                                           partition_start, partition_end)
 
-            if fs_type is not None:
+            if mkfs_in_dom0:
+                session.call_plugin_serialized('partition_utils.py', 'mkfs',
+                                               dev, '1', fs_type, fs_label)
+
+        # 3.a. dom0 does not support nfs/ext4, so may have to mkfs in domU
+        if fs_type is not None and not mkfs_in_dom0:
+            with vdi_attached(session, vdi_ref, read_only=False) as dev:
+                partition_path = utils.make_dev_path(dev, partition=1)
                 utils.mkfs(fs_type, partition_path, fs_label,
                            run_as_root=True)
 
@@ -1009,6 +1034,11 @@ def generate_swap(session, instance, vm_ref, userdevice, name_label, swap_mb):
     # partition because that is what parted supports.
     is_windows = instance['os_type'] == "windows"
     fs_type = "vfat" if is_windows else "swap"
+
+    if CONF.xenserver.independent_compute and fs_type != "swap":
+        raise exception.NotSupportedWithOption(
+            operation='swap drives for Windows',
+            option='CONF.xenserver.independent_compute')
 
     _generate_disk(session, instance, vm_ref, userdevice, name_label,
                    'swap', swap_mb, fs_type)
@@ -1082,28 +1112,26 @@ def generate_configdrive(session, instance, vm_ref, userdevice,
     sr_ref = safe_find_sr(session)
     vdi_ref = create_vdi(session, sr_ref, instance, 'config-2',
                          'configdrive', configdrive.CONFIGDRIVESIZE_BYTES)
-
     try:
-        with vdi_attached_here(session, vdi_ref, read_only=False) as dev:
-            extra_md = {}
-            if admin_password:
-                extra_md['admin_pass'] = admin_password
-            inst_md = instance_metadata.InstanceMetadata(instance,
-                    content=files, extra_md=extra_md,
-                    network_info=network_info)
-            with configdrive.ConfigDriveBuilder(instance_md=inst_md) as cdb:
-                with utils.tempdir() as tmp_path:
-                    tmp_file = os.path.join(tmp_path, 'configdrive')
-                    cdb.make_drive(tmp_file)
-
-                    dev_path = utils.make_dev_path(dev)
-                    utils.execute('dd',
-                                  'if=%s' % tmp_file,
-                                  'of=%s' % dev_path,
-                                  'bs=%d' % DD_BLOCKSIZE,
-                                  'oflag=direct,sync',
-                                  run_as_root=True)
-
+        extra_md = {}
+        if admin_password:
+            extra_md['admin_pass'] = admin_password
+        inst_md = instance_metadata.InstanceMetadata(
+            instance, content=files, extra_md=extra_md,
+            network_info=network_info)
+        with configdrive.ConfigDriveBuilder(instance_md=inst_md) as cdb:
+            with utils.tempdir() as tmp_path:
+                tmp_file = os.path.join(tmp_path, 'configdrive')
+                cdb.make_drive(tmp_file)
+                # XAPI can only import a VHD file, so convert to vhd format
+                vhd_file = '%s.vhd' % tmp_file
+                utils.execute('qemu-img', 'convert', '-Ovpc', tmp_file,
+                              vhd_file)
+                vhd_file_size = os.path.getsize(vhd_file)
+                with open(vhd_file) as file_obj:
+                    volume_utils.stream_to_vdi(
+                        session, instance, 'vhd', file_obj,
+                        vhd_file_size, vdi_ref)
         create_vbd(session, vm_ref, vdi_ref, userdevice, bootable=False,
                    read_only=True)
     except Exception:
@@ -1120,6 +1148,11 @@ def _create_kernel_image(context, session, instance, name_label, image_id,
 
     Returns: A list of dictionaries that describe VDIs
     """
+    if CONF.xenserver.independent_compute:
+        raise exception.NotSupportedWithOption(
+            operation='Non-VHD images',
+            option='CONF.xenserver.independent_compute')
+
     filename = ""
     if CONF.xenserver.cache_images != 'none':
         args = {}
@@ -1289,6 +1322,10 @@ def _fetch_image(context, session, instance, name_label, image_id, image_type):
     if image_type == ImageType.DISK_VHD:
         vdis = _fetch_vhd_image(context, session, instance, image_id)
     else:
+        if CONF.xenserver.independent_compute:
+            raise exception.NotSupportedWithOption(
+                operation='Non-VHD images',
+                option='CONF.xenserver.independent_compute')
         vdis = _fetch_disk_image(context, session, instance, name_label,
                                  image_id, image_type)
 
@@ -1496,7 +1533,7 @@ def _fetch_disk_image(context, session, instance, name_label, image_id,
         filename = None
         vdi_uuid = session.call_xenapi("VDI.get_uuid", vdi_ref)
 
-        with vdi_attached_here(session, vdi_ref, read_only=False) as dev:
+        with vdi_attached(session, vdi_ref, read_only=False) as dev:
             _stream_disk(
                 session, image.stream_to, image_type, virtual_size, dev)
 
@@ -1658,7 +1695,7 @@ def preconfigure_instance(session, instance, vdi_ref, network_info):
     if not mount_required:
         return
 
-    with vdi_attached_here(session, vdi_ref, read_only=False) as dev:
+    with vdi_attached(session, vdi_ref, read_only=False) as dev:
         _mounted_processing(dev, key, net, metadata)
 
 
@@ -2080,16 +2117,24 @@ def _remap_vbd_dev(dev):
     return remapped_dev
 
 
-def _wait_for_device(dev):
+def _wait_for_device(session, dev, dom0, max_seconds):
     """Wait for device node to appear."""
-    for i in range(0, CONF.xenserver.block_device_creation_timeout):
-        dev_path = utils.make_dev_path(dev)
-        if os.path.exists(dev_path):
-            return
-        time.sleep(1)
+    dev_path = utils.make_dev_path(dev)
+    found_path = None
+    if dom0:
+        found_path = session.call_plugin_serialized('partition_utils.py',
+                                                    'wait_for_dev',
+                                                    dev_path, max_seconds)
+    else:
+        for i in range(0, max_seconds):
+            if os.path.exists(dev_path):
+                found_path = dev_path
+                break
+            time.sleep(1)
 
-    raise exception.StorageError(
-        reason=_('Timeout waiting for device %s to be created') % dev)
+    if found_path is None:
+        raise exception.StorageError(
+            reason=_('Timeout waiting for device %s to be created') % dev)
 
 
 def cleanup_attached_vdis(session):
@@ -2116,8 +2161,13 @@ def cleanup_attached_vdis(session):
 
 
 @contextlib.contextmanager
-def vdi_attached_here(session, vdi_ref, read_only=False):
-    this_vm_ref = _get_this_vm_ref(session)
+def vdi_attached(session, vdi_ref, read_only=False, dom0=False):
+    if dom0:
+        this_vm_ref = _get_dom0_ref(session)
+    else:
+        # Make sure we are running as a domU.
+        ensure_correct_host(session)
+        this_vm_ref = _get_this_vm_ref(session)
 
     vbd_ref = create_vbd(session, this_vm_ref, vdi_ref, 'autodetect',
                          read_only=read_only, bootable=False)
@@ -2134,10 +2184,13 @@ def vdi_attached_here(session, vdi_ref, read_only=False):
                 LOG.debug('VBD %(vbd_ref)s plugged into wrong dev, '
                           'remapping to %(dev)s',
                           {'vbd_ref': vbd_ref, 'dev': dev})
-            _wait_for_device(dev)
+            _wait_for_device(session, dev, dom0,
+                             CONF.xenserver.block_device_creation_timeout)
             yield dev
         finally:
-            utils.execute('sync', run_as_root=True)
+            # As we can not have filesystems mounted here (we cannot
+            # destroy the VBD with filesystems mounted), it is not
+            # useful to call sync.
             LOG.debug('Destroying VBD for VDI %s ... ', vdi_ref)
             unplug_vbd(session, vbd_ref, this_vm_ref)
     finally:
@@ -2154,11 +2207,29 @@ def _get_sys_hypervisor_uuid():
         return f.readline().strip()
 
 
+def _get_dom0_ref(session):
+    vms = session.call_xenapi("VM.get_all_records_where",
+                              'field "domid"="0" and '
+                              'field "resident_on"="%s"' %
+                              session.host_ref)
+    return list(vms.keys())[0]
+
+
 def get_this_vm_uuid(session):
+    if CONF.xenserver.independent_compute:
+        msg = _LE("This host has been configured with the independent "
+                  "compute flag.  An operation has been attempted which is "
+                  "incompatible with this flag, but should have been "
+                  "caught earlier.  Please raise a bug against the "
+                  "OpenStack Nova project")
+        LOG.error(msg)
+        raise exception.NotSupportedWithOption(
+            operation='uncaught operation',
+            option='CONF.xenserver.independent_compute')
     if session and session.is_local_connection:
         # UUID is the control domain running on this host
         vms = session.call_xenapi("VM.get_all_records_where",
-                                  'field "is_control_domain"="true" and '
+                                  'field "domid"="0" and '
                                   'field "resident_on"="%s"' %
                                   session.host_ref)
         return vms[list(vms.keys())[0]]['uuid']
@@ -2353,10 +2424,10 @@ def _copy_partition(session, src_ref, dst_ref, partition, virtual_size):
     # Part of disk taken up by MBR
     virtual_size -= MBR_SIZE_BYTES
 
-    with vdi_attached_here(session, src_ref, read_only=True) as src:
+    with vdi_attached(session, src_ref, read_only=True) as src:
         src_path = utils.make_dev_path(src, partition=partition)
 
-        with vdi_attached_here(session, dst_ref, read_only=False) as dst:
+        with vdi_attached(session, dst_ref, read_only=False) as dst:
             dst_path = utils.make_dev_path(dst, partition=partition)
 
             _write_partition(session, virtual_size, dst)
@@ -2421,12 +2492,17 @@ def _mounted_processing(device, key, net, metadata):
 
 def ensure_correct_host(session):
     """Ensure we're connected to the host we're running on. This is the
-    required configuration for anything that uses vdi_attached_here.
+    required configuration for anything that uses vdi_attached without
+    the dom0 flag.
     """
+    if session.host_checked:
+        return
+
     this_vm_uuid = get_this_vm_uuid(session)
 
     try:
         session.call_xenapi('VM.get_by_uuid', this_vm_uuid)
+        session.host_checked = True
     except session.XenAPI.Failure as exc:
         if exc.details[0] != 'UUID_INVALID':
             raise
